script: Move htmlinputelement into `script/dom/html/form_controls`
It also moves the input types into their dedicated folder `input_type/`

Part of #46329

Testing: it compiles
