+++ components/script/dom/event/event.rs
@@ -38,8 +38,8 @@ use crate::dom::bindings::str::DOMString;
 use crate::dom::element::Element;
 use crate::dom::eventtarget::{EventListeners, EventTarget, ListenerPhase};
 use crate::dom::globalscope::GlobalScope;
+use crate::dom::html::form_controls::htmlinputelement::InputActivationState;
 use crate::dom::html::htmlslotelement::HTMLSlotElement;
-use crate::dom::html::input_element::InputActivationState;
 use crate::dom::mouseevent::MouseEvent;
 use crate::dom::node::virtualmethods::vtable_for;
 use crate::dom::node::{Node, NodeTraits};
