+++ components/script/dom/html/form_controls/input_type/hidden_input_type.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use crate::dom::input_element::input_type::SpecificInputType;
+use crate::dom::html::form_controls::input_type::SpecificInputType;
 
 #[derive(Default, JSTraceable, MallocSizeOf, PartialEq)]
 pub(crate) struct HiddenInputType();
