+++ components/script/dom/html/form_controls/input_type/month_input_type.rs
@@ -8,9 +8,9 @@ use script_bindings::cell::DomRefCell;
 use time::{Month, OffsetDateTime};
 
 use crate::dom::bindings::str::{DOMString, FromInputValueString, ToInputValueString};
-use crate::dom::htmlinputelement::text_input_widget::TextInputWidget;
-use crate::dom::input_element::HTMLInputElement;
-use crate::dom::input_element::input_type::SpecificInputType;
+use crate::dom::html::form_controls::htmlinputelement::HTMLInputElement;
+use crate::dom::html::form_controls::input_type::SpecificInputType;
+use crate::dom::html::form_controls::input_type::text_input_widget::TextInputWidget;
 
 #[derive(Default, JSTraceable, MallocSizeOf, PartialEq)]
 #[cfg_attr(crown, crown::unrooted_must_root_lint::must_root)]
