+++ components/script/dom/html/form_controls/input_type/password_input_type.rs
@@ -3,28 +3,21 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 use js::context::JSContext;
 use script_bindings::cell::DomRefCell;
-use url::Url;
 
 use crate::dom::bindings::str::DOMString;
-use crate::dom::htmlinputelement::text_input_widget::TextInputWidget;
-use crate::dom::input_element::HTMLInputElement;
-use crate::dom::input_element::input_type::SpecificInputType;
+use crate::dom::html::form_controls::htmlinputelement::HTMLInputElement;
+use crate::dom::html::form_controls::input_type::SpecificInputType;
+use crate::dom::html::form_controls::input_type::text_input_widget::TextInputWidget;
 
 #[derive(Default, JSTraceable, MallocSizeOf, PartialEq)]
 #[cfg_attr(crown, crown::unrooted_must_root_lint::must_root)]
-pub(crate) struct UrlInputType {
+pub(crate) struct PasswordInputType {
     text_input_widget: DomRefCell<TextInputWidget>,
 }
 
-impl SpecificInputType for UrlInputType {
+impl SpecificInputType for PasswordInputType {
     fn sanitize_value(&self, _input: &HTMLInputElement, value: &mut DOMString) {
         value.strip_newlines();
-        value.strip_leading_and_trailing_ascii_whitespace();
-    }
-
-    /// <https://html.spec.whatwg.org/multipage/#url-state-(type=url):suffering-from-a-type-mismatch>
-    fn suffers_from_type_mismatch(&self, _input: &HTMLInputElement, value: &DOMString) -> bool {
-        Url::parse(&value.str()).is_err()
     }
 
     fn update_shadow_tree(&self, cx: &mut JSContext, input: &HTMLInputElement) {
