+++ components/script/dom/html/form_controls/mod.rs
@@ -2,5 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+pub(crate) mod htmlinputelement;
+pub(crate) mod input_type;
 pub(crate) mod text_control;
 pub(crate) mod text_input;
